% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{add_gene_to_DA_list}
\alias{add_gene_to_DA_list}
\title{Add genes to differential analysis list using a SingleCellExperiment}
\usage{
add_gene_to_DA_list(
  scExp,
  IDC_DA_list,
  feature_ID_col = "ID",
  gene_col = "Gene",
  distanceToTSS = 1000,
  split = TRUE,
  split_char = ", "
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object. The rowData must contain a
column with genes (see param 'gene_col').}

\item{IDC_DA_list}{A list of data.frame of  differential features at each
clustering  iteration produced by \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}
("IDC_DA.qs").}

\item{feature_ID_col}{A character specifying the column in which to retrieve
the feature ID.}

\item{gene_col}{A character specifying the column in which to retrieve the
gene / feature name.}

\item{distanceToTSS}{An integer specifying the maximum distance to consider
a feature close enough to a gene.}

\item{split}{A logical indicating wether to split the gene column or not}

\item{split_char}{If split is TRUE. A character by which the gene column is
splitted.}
}
\value{

}
\description{
Add genes to differential analysis list using a SingleCellExperiment
}
\examples{
data("scExp", package = "IDclust")
data("IDC_DA_scEpigenomics", package = "IDclust")

IDC_DA_scEpigenomics = add_gene_to_DA_list(
    scExp = scExp, 
    IDC_DA_list = IDC_DA_scEpigenomics,
    feature_ID_col = "ID", 
    gene_col = "Gene", 
    distanceToTSS = 1000,
    split = TRUE,
    split_char = ", "
)

}
