% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_positive.R
\name{calculate_FalseNegative}
\alias{calculate_FalseNegative}
\title{Calculate False Negative markers for SingleCellExperiment object}
\usage{
calculate_FalseNegative(
  object,
  true_marker_list,
  differential_function = differential_ChromSCape,
  by = "IDcluster",
  logFC.th = log2(1.5),
  qval.th = 0.01,
  ...
)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{differential_function}{A function that take in entry a
SingleCellExperiment object and  parameters passed in ... and returns a
data.frame containing the significantly differential features for each
cluster.}

\item{by}{A character specifying the name of the metadata column referencing
the clusters.}

\item{...}{Additional parameters passed to the differential_function. See
\code{\link[=differential_ChromSCape]{differential_ChromSCape()}} for more information on additional
parameters for the default function.}

\item{nThreads}{If runFDR==TRUE. An integer specifying of threads to use
for the calculation of the FDR.}

\item{iterations}{An integer specifyung the number of iterations of random
subsampling to run.}
}
\description{
Calculate False Negative markers for SingleCellExperiment object
}
\examples{
if(requireNamespace("ChromSCape", quietly=TRUE)){
data("scExp", package = "IDclust")
scExp = ChromSCape::find_clusters_louvain_scExp(scExp, resolution = 0.1)
FDR_df = calculate_FalseNegative(scExp, nThreads = 1, iterations = 2)
head(FDR_df)
}
}
