% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_pseudobulk_mat}
\alias{create_pseudobulk_mat}
\title{Create Pseudo-Bulk Matrice from scRNA clusters & replicates}
\usage{
create_pseudobulk_mat(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Create Pseudo-Bulk Matrice from scRNA clusters & replicates
}
\examples{
if(requireNamespace("Seurat", quietly=TRUE)){
data("Seu", package = "IDclust")
mat <- create_pseudobulk_mat(Seu, by = "seurat_clusters")
}

#' if(requireNamespace("SingleCellExperiment", quietly=TRUE)){
data("scExp", package = "IDclust")
mat <- create_pseudobulk_mat(scExp, by = "cell_clusters")
}
}
