% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_clustering_scRNA_Seurat.R
\name{create_pseudobulk_mat_Seu}
\alias{create_pseudobulk_mat_Seu}
\title{Create Pseudo-Bulk Matrice from scRNA clusters & replicates}
\usage{
create_pseudobulk_mat_Seu(Seu, biological_replicate_col)
}
\arguments{
\item{Seu}{A Seurat object containing scRNA dataset with 'seurat_clusters'
column.}

\item{biological_replicate_col}{Optional. A column of the Seurat object
indicating the replicates or batches of the dataset in order to take in
account biological/technical noise. If NULL, will create random layers of
fake replicates.}
}
\value{
A pseudo-bulk matrice of cluster spread by replicates / batches /
fake replicates.
}
\description{
Create Pseudo-Bulk Matrice from scRNA clusters & replicates
}
