% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{differential_ChromSCape}
\alias{differential_ChromSCape}
\title{Differential analysis with ChromSCape (differential activation)}
\usage{
differential_ChromSCape(
  object,
  by = "IDcluster",
  logFC.th = log2(2),
  qval.th = 0.01,
  min.pct = 0.2,
  marker = c("pos", "neg", "diff")[1]
)
}
\arguments{
\item{object}{A SingleCellExperiment object containing scRNA dataset with a
metadata column name matching the by parameter}

\item{by}{A character specifying the name of the metadata column referencing
the clusters.}

\item{logFC.th}{A positive numeric specifying the log2 fold change of activation
above/below which a feature is considered as significantly differential.}

\item{qval.th}{A numeric specifying the adjusted p-value below
which a feature is considered as significantly differential.}

\item{min.pct}{Minimum percentage of cells to be active in the cells of the
cluster to consider a feature as potentially significantly differential.}

\item{marker}{A character specifying which markers to retrieve. 'pos' for
positive markers (default), 'neg' for negative markers or 'diff' for
differential markers.}
}
\description{
Runs \code{\link[ChromSCape:differential_activation]{ChromSCape::differential_activation()}} function on clusters in the
'by' column of the SingleCellExperiment object and returns a data.frame
containing the differential features that passed the thresholds.
}
\examples{
if(requireNamespace("ChromSCape", quietly=TRUE)){
 data("scExp", package = "IDclust")
DA <- differential_ChromSCape(scExp)
}
}
\seealso{
See \code{\link[ChromSCape:differential_activation]{ChromSCape::differential_activation()}}
}
