% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{differential_edgeR_pseudobulk_LRT.default}
\alias{differential_edgeR_pseudobulk_LRT.default}
\title{Pseudo-bulk differential analysis with edgeR (LRT)}
\usage{
\method{differential_edgeR_pseudobulk_LRT}{default}(
  object,
  by = "IDcluster",
  biological_replicate_col = NULL,
  logFC.th = log2(2),
  qval.th = 0.01,
  min.pct = 0.2
)
}
\arguments{
\item{object}{A SingleCellExperiment object containing scRNA dataset with a
metadata column name matching the by parameter}

\item{by}{A character specifying the name of the metadata column referencing
the clusters.}

\item{biological_replicate_col}{A character specifying the column of the
object metadata definin the biological / technical replicates. If NULL, three
random set of replicates will be created per cluster, provided there are
enough cells.}

\item{logFC.th}{A numeric specifying the log2 fold change of activation
above/below which a feature is considered as significantly differential.}

\item{qval.th}{A numeric specifying the adjusted p-value below
which a feature is considered as significantly differential.}

\item{min.pct}{Minimum percentage of cells to be active in the cells of the
cluster to consider a feature as potentially significantly differential.}
}
\value{
A data.frame containing the results of the differential analysis
}
\description{
Pseudo-bulk differential analysis with edgeR (LRT)
}
\details{
Concatenate single-cells into replicates by cluster in order to
create a 'pseudo-bulk' matrice of multiple replicates per cluster. If no
replicates are present, will assign replicates at random to create 3 replicates
per cluster. Conducts 'LRT' (likelihood ratio tests) edgeR tests to test.
See \code{\link[edgeR:glmfit]{edgeR::glmLRT()}}
}
\examples{
if(requireNamespace("Seurat", quietly=TRUE)){
data("Seu", package = "IDclust")
DA <- differential_edgeR_pseudobulk_LRT(Seu)
}
}
\seealso{
See \code{\link[edgeR:glmfit]{edgeR::glmLRT()}}
}
