% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{differential_edgeR_pseudobulk_LRT}
\alias{differential_edgeR_pseudobulk_LRT}
\title{Pseudo-bulk differential analysis with edgeR (LRT)}
\usage{
differential_edgeR_pseudobulk_LRT(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Pseudo-bulk differential analysis with edgeR (LRT)
}
\examples{
if(requireNamespace("Seurat", quietly=TRUE)){
data("Seu", package = "IDclust")
DA <- differential_edgeR_pseudobulk_LRT(Seu)
}
if(requireNamespace("SingleCellExperiment", quietly=TRUE)){
data("scExp", package = "IDclust")
DA <- differential_edgeR_pseudobulk_LRT(scExp)
}
}
