% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{differential_pathway}
\alias{differential_pathway}
\title{Find differential pathways with enrichR}
\usage{
differential_pathway(
  genes,
  qval.th = 0.1,
  website = c("Enrichr", "FlyEnrichr", "WormEnrichr", "YeastEnrichr", "FishEnrichr")[1],
  databases = c("MSigDB_Hallmark_2020", "GO_Molecular_Function_2021",
    "MSigDB_Oncogenic_Signatures"),
  order_by_database = TRUE
)
}
\arguments{
\item{genes}{A character vector containing at least 10 genes}

\item{qval.th}{A numeric specifying the adjusted p.value below which a
pathway is considered as significantly enriched}

\item{website}{site requested}

\item{databases}{(Required). Character vector of databases to search.
See https://maayanlab.cloud/Enrichr/ for available databases.}

\item{order_by_database}{A logical. If TRUE, row will appear in the order of
the databases vector, then within each database are sorted by adjusted
pvalue.}
}
\value{
A dataframe with enriched pathways.
}
\description{
Find differential pathways with enrichR
}
\examples{


pathways <- differential_pathway(
c("Runx1", "Gfi1", "Gfi1b", "Spi1", "Gata1", "Kdr"))

}
\references{
Chen, E.Y., Tan, C.M., Kou, Y. et al. Enrichr: interactive and
collaborative HTML5 gene list enrichment analysis tool. BMC Bioinformatics 14,
128 (2013). https://doi.org/10.1186/1471-2105-14-128
}
