% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{differential_Seurat}
\alias{differential_Seurat}
\title{Differential function Seurat (wilcoxon)}
\usage{
differential_Seurat(
  object,
  by = "cell_cluster",
  logFC.th = log2(1.5),
  qval.th = 0.01,
  min.pct = 0.1,
  ...
)
}
\arguments{
\item{object}{A Seurat object containing scRNA dataset with a metadata column
name matching the by parameter}

\item{by}{A character specifying the name of the metadata column referencing
the clusters.}

\item{logFC.th}{A numeric specifying the log2 fold change of activation
above/below which a feature is considered as significantly differential.}

\item{qval.th}{A numeric specifying the adjusted p-value below
which a feature is considered as significantly differential.}

\item{min.pct}{Minimum percentage of cells to be active in the cells of the
cluster to consider a feature as potentially significantly differential.}

\item{...}{Additional parameters to pass to \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}.}
}
\value{

}
\description{
Run \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}  function on the
clusters in the 'by' column and return a data.frame  containing the marker
genes of each cluster passing the thresholds .
}
\examples{
if(requireNamespace("Seurat", quietly=TRUE)){
data("Seu", package = "IDclust")
DA <- differential_Seurat(Seu)
}
}
\seealso{
See \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}
}
