% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IDC_DA_scEpigenomics}
\alias{IDC_DA_scEpigenomics}
\title{A list of data.frame containing the differential markers for scEpigenomics dataset
(Paired-Tag - scEpigenomics)}
\format{
IDC_DA_scEpigenomics - A data.frame obtained by
\code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}.
}
\usage{
data("IDC_DA_scEpigenomics")
}
\description{
A list of differential markers for scEpigenomics
}
\references{
Joint profiling of histone modifications and transcriptome in single cells
from mouse brain, Chenxu Zhu, Yanxiao Zhang,  Yang Eric Li, Jacinta Lucero,
. Margarita Behrens, Bing Ren, Nature Methods, 2021
\href{https://doi.org/10.1038/s41592-021-01060-3}{Paired-Tag}
}
\keyword{datasets}
