% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IDC_summary_scEpigenomics}
\alias{IDC_summary_scEpigenomics}
\title{A data.frame containing the summary of differential markers for scEpigenomics dataset
(Paired-Tag - H3K27ac)}
\format{
IDC_summary_scEpigenomics - A data.frame obtained by
\code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}.
}
\usage{
data("IDC_summary_scEpigenomics")
}
\description{
A summary of differential markers for scEpigenomics (Paired-Tag - H3K27ac)
}
\references{
Joint profiling of histone modifications and transcriptome in single cells
from mouse brain, Chenxu Zhu, Yanxiao Zhang,  Yang Eric Li, Jacinta Lucero,
. Margarita Behrens, Bing Ren, Nature Methods, 2021
\href{https://doi.org/10.1038/s41592-021-01060-3}{Paired-Tag}
}
\keyword{datasets}
