% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_clustering_scHistone.R
\name{iterative_differential_clustering_scEpigenomics}
\alias{iterative_differential_clustering_scEpigenomics}
\title{Iterative Differential Clustering scEpigenomics}
\usage{
iterative_differential_clustering_scEpigenomics(
  scExp,
  output_dir = "./",
  plotting = TRUE,
  saving = TRUE,
  nPCA = 10,
  percent_feature = 1,
  quantile.activation = 0.7,
  min.pct.cell_assigned = 0.1,
  FC.th = 2,
  qval.th = 0.1,
  limit = 5,
  k = 100,
  starting.resolution = 0.1,
  resolution = 0.8,
  runFDR = FALSE,
  limit_by_proportion = NULL,
  color = NULL,
  nThreads = 10,
  verbose = TRUE
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object}

\item{output_dir}{}

\item{plotting}{}

\item{nPCA}{}

\item{percent_feature}{}

\item{quantile.activation}{}

\item{min.pct.cell_assigned}{}

\item{FC.th}{}

\item{qval.th}{}

\item{limit}{}

\item{k}{}

\item{resolution}{}

\item{runFDR}{}

\item{limit_by_proportion}{}

\item{nThreads}{}
}
\value{

}
\description{
Iterative Differential Clustering scEpigenomics
}
\examples{
library(ChromSCape)
data("scExp")
output_dir = tempdir()
iterative_differential_clustering_scEpigenomics(
scExp,
output_dir = "./",
plotting = TRUE,
saving = TRUE,
nPCA = 10,
percent_feature = 1,
quantile.activation = 0.7,
min.pct.cell_assigned = 0.1,
FC.th = 2,
qval.th = 0.1,
limit = 5,
k = 100,
starting.resolution = 0.1,
resolution = 0.8,
runFDR = FALSE,
limit_by_proportion = NULL,
color = NULL,
nThreads = 10,
verbose = TRUE
)

}
