% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_network}
\alias{plot_cluster_network}
\alias{plot_cluster_network.default}
\alias{plot_cluster_network.Seurat}
\title{Plot Iterative Differential Clustering network}
\usage{
plot_cluster_network(object, ...)

\method{plot_cluster_network}{default}(
  object,
  differential_summary_df,
  color_by = "cell_cluster",
  cluster_col = "cell_cluster",
  colors = NULL,
  node_size_factor = 7.5,
  edge_size_factor = 1,
  function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
    flip.y = F),
  ...
)

\method{plot_cluster_network}{Seurat}(
  object,
  differential_summary_df,
  color_by = "cell_cluster",
  cluster_col = "cell_cluster",
  colors = NULL,
  node_size_factor = 7.5,
  edge_size_factor = 1,
  function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
    flip.y = F),
  ...
)
}
\arguments{
\item{object}{A Seurat object clustered with \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}}

\item{...}{Additional parameters passed to the plot function.}

\item{differential_summary_df}{Optional. A data.frame of differential
analyses summary outputed by \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}} when
saving option is TRUE. Use to determine the width of the edges based on the
number of differential features of the given marker.}

\item{color_by}{A character specifying the column of the Seurat
to use for coloring the nodes.}

\item{cluster_col}{A character specifying the column of the
Seurat to use to store the iterative differential clusters.}

\item{node_size_factor}{A numeric specifying the size of the nodes.}

\item{function_layout}{A function of g for the layout of the graph.}
}
\value{


A hierarchical network of cluster assignation:
\itemize{
\item Size of nodes reflects the number of cells
\item Width of edges reflects the number of differential features defining a cluster
\item Color of nodes reflects the repartition of cells according to 'color_by'
}

A hierarchical network of cluster assignation:
\itemize{
\item Size of nodes reflects the number of cells
\item Width of edges reflects the number of differential features defining a cluster
\item Color of nodes reflects the repartition of cells according to 'color_by'
}
}
\description{
Plot Iterative Differential Clustering network

Plot Iterative Differential Clustering network

Plot Iterative Differential Clustering network
}
