% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processing_ChromSCape}
\alias{processing_ChromSCape}
\title{Normalize and run PCA with ChromSCape (LSI)}
\usage{
processing_ChromSCape(scExp, n_dims = 10, dim_red = "PCA")
}
\arguments{
\item{scExp}{A Seurat object to preprocess with Seurat.}

\item{n_dims}{An integer specifying the number of first dimensions to keep
in the dimensionality reduction step.}

\item{dim_red}{The name of the slot to save the dimensionality reduction at
each step in the  \code{reducedDimNames(scExp)}.}
}
\value{
A SingleCellExperiment object with the dimensionality reduction filled
}
\description{
Runs ChromSCape LSI and  removes the first
component, often driven by total library size.
See \code{\link[ChromSCape:preprocessing_filtering_and_reduction]{ChromSCape::preprocessing_filtering_and_reduction()}}.
}
\examples{
data("scExp", package = "IDclust")
processing_ChromSCape(scExp)
}
