% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processing_Seurat}
\alias{processing_Seurat}
\title{Normalize and run PCA with Seurat}
\usage{
processing_Seurat(Seu, n_dims = 50, dim_red = "pca")
}
\arguments{
\item{Seu}{A Seurat object to preprocess with Seurat.}

\item{n_dims}{An integer specifying the number of first dimensions to keep
in the dimensionality reduction step.}

\item{dim_red}{The name of the slot to save the dimensionality reduction at
each step in the  \code{Seurat::Reductions(Seu)}.}
}
\value{
A Seurat object with the dimensionality reduction filled
}
\description{
Runs Seurat PCA with 10 PC and by removing the first
component. See \code{\link[ChromSCape:preprocessing_filtering_and_reduction]{ChromSCape::preprocessing_filtering_and_reduction()}}
}
