% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_clustering_scRNA_Seurat.R
\name{pseudobulk_LRT_edgeR}
\alias{pseudobulk_LRT_edgeR}
\title{Pseudo-bulk differential analysis (LRT) edgeR}
\usage{
pseudobulk_LRT_edgeR(Seu, biological_replicate_col = NULL)
}
\arguments{
\item{Seu}{A Seurat object containing scRNA dataset with 'seurat_clusters'
column}

\item{biological_replicate_col}{Optional. A column of the Seurat object
indicating the replicates or batches of the dataset in order to take in
account biological/technical noise. If NULL, will create random layers of
fake replicates.}
}
\value{
A data.frame containing the results of the differential analysis
}
\description{
Pseudo-bulk differential analysis (LRT) edgeR
}
\details{
Concatenate single-cells into replicates by cluster in order to
create a 'pseudo-bulk' matrice of multiple replicates per cluster. If no
replicates are present, will assign replicates at random to create 3 replicates
per cluster. Conducts 'LRT' (likelihood ratio tests) edgeR tests to test
}
\seealso{
glmFit {edgeR}
}
