% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scExp}
\alias{scExp}
\title{A SingleCellExperiment object containing the scH3K27ac part of Paired-Tag
dataset. (Paired-Tag - H3K27ac)}
\format{
scExp - A SingleCellExperiment object. See
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}.
}
\usage{
data("scExp")
}
\description{
A SingleCellExperiment object of a  subset of 2000 x 5000 cells
by features of the
\href{https://doi.org/10.1038/s41592-021-01060-3}{Paired-Tag} multiomic
paper. The metadata contains annotation from the author as well as
iterative differential clusters run on the subset of cells and features The
counts were analyzed with ChromSCape package.
}
\references{
Joint profiling of histone modifications and transcriptome in single cells
from mouse brain, Chenxu Zhu, Yanxiao Zhang,  Yang Eric Li, Jacinta Lucero,
. Margarita Behrens, Bing Ren, Nature Methods, 2021
\href{https://doi.org/10.1038/s41592-021-01060-3}{Paired-Tag}
}
\keyword{datasets}
