% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seu}
\alias{Seu}
\title{A Seurat object containing the scRNA part of Paired-Tag dataset. (Paired-Tag - scRNA)}
\format{
Seu - A Seurat object. See \code{\link[Seurat:Seurat-package]{Seurat::Seurat()}}.
}
\usage{
data("Seu")
}
\description{
A Seurat object of a  subset of 2000 x 2000 cells by genes of
the \href{https://doi.org/10.1038/s41592-021-01060-3}{Paired-Tag} multiomic
paper. The metadata contains annotation from the author as well as
iterative differential clusters run on the subset of cells and genes. The
counts were analyzed with Seurat package.
}
\references{
Joint profiling of histone modifications and transcriptome in single cells
from mouse brain, Chenxu Zhu, Yanxiao Zhang,  Yang Eric Li, Jacinta Lucero,
. Margarita Behrens, Bing Ren, Nature Methods, 2021
\href{https://doi.org/10.1038/s41592-021-01060-3}{Paired-Tag}
}
\keyword{datasets}
