% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarise_DA}
\alias{summarise_DA}
\title{Summarise Differential Analysis table}
\usage{
summarise_DA(
  res,
  chr_col = "chr",
  start_col = "start",
  end_col = "end",
  ID_col = "ID"
)
}
\arguments{
\item{res}{A differential analysis data.frame}

\item{chr_col}{Character specifying the chromosome column.}

\item{start_col}{Character specifying the start column.}

\item{end_col}{Character specifying the end column.}

\item{ID_col}{Character specifying the ID column.}
}
\value{
A gathered data.frame with a cluster column
}
\description{
Summarise Differential Analysis table
}
