% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{top_differential_markers}
\alias{top_differential_markers}
\title{Retrieve top marker genes from list of differential analysis}
\usage{
top_differential_markers(
  object,
  top = 1,
  gene_col = "gene",
  logFC_col = "avg_log2FC",
  qvalue_col = "p_val_adj",
  order_by = c("logFC_col", "qvalue_col")[1],
  pseudogene_pattern = NULL
)
}
\arguments{
\item{object}{A list of data.frame of  differential features at each
clustering  iteration produced by \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}
("IDC_DA.qs").}

\item{top}{An integer specifying the number of top features to retrieve per
cluster.}

\item{gene_col}{A character specifying the column in which to retrieve the
gene / feature name.}

\item{logFC_col}{A character specifying the column in which to retrieve the
logFC.}

\item{qvalue_col}{A character specifying the column in which to retrieve the
adjusted p.value.}

\item{order_by}{A character specifying the column by which to order the top
markers (default to logFC_col.}

\item{pseudogene_pattern}{A character specifying the pattern of 'pseudo-genes'
to exclude from the top markers.}
}
\value{

}
\description{
Retrieve top marker genes from list of differential analysis
}
\examples{
#scRNA
data("IDC_DA_scRNA", package = "IDclust")

top_differential_markers(
    IDC_DA_scRNA,
    top = 1,
    gene_col = "gene",
    logFC_col = "avg_log2FC",
    qvalue_col = "p_val_adj",
    order_by = "logFC_col",
    pseudogene_pattern = NULL
)

#scEpigenomics
data("scExp", package = "IDclust")
data("IDC_DA_scEpigenomics", package = "IDclust")

# We must first add the gene information to the DA list: 
IDC_DA_scEpigenomics = add_gene_to_DA_list(
    scExp = scExp, 
    IDC_DA_list = IDC_DA_scEpigenomics,
    feature_ID_col = "ID", 
    gene_col = "Gene", 
    distanceToTSS = 1000,
    split = TRUE,
    split_char = ", "
)

top_differential_markers(
    IDC_DA_scEpigenomics,
    top = 1,
    gene_col = "Gene",
    logFC_col = "logFC",
    qvalue_col = "qval",
    order_by = "logFC_col",
    pseudogene_pattern = "Rik|Vmn|Gm|AW"
)
    
}
