% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{top_differential_markers.SingleCellExperiment}
\alias{top_differential_markers.SingleCellExperiment}
\title{Retrieve Top differential markers}
\usage{
top_differential_markers.SingleCellExperiment(
  scExp,
  IDC_DA_list,
  top = 1,
  gene_col = "Gene",
  pseudogene_pattern = "Rik|Vmn|Gm|AW",
  distanceToTSS = 1000
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object}

\item{IDC_DA_list}{A list of data.frame of  differential features at each
clustering  iteration produced by \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}
("IDC_DA.qs").}

\item{top}{An integer specifying the number of top features to retrieve per
cluster.}

\item{gene_col}{A character specifying the column in which to retrieve the
gene / feature name.}

\item{pseudogene_pattern}{A character specifying the pattern of 'pseudo-genes'
to exclude from the top markers.}

\item{distanceToTSS}{An integer specifying the maximum distance to consider
a feature close enough to a gene.}
}
\value{

}
\description{
Retrieve Top differential markers
}
