% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_pseudobulk_mat_Seu}
\alias{create_pseudobulk_mat_Seu}
\title{Create Pseudo-Bulk Matrice from scRNA clusters & replicates}
\usage{
create_pseudobulk_mat_Seu(
  Seu,
  by = "cell_cluster",
  biological_replicate_col = NULL,
  assay = "RNA"
)
}
\arguments{
\item{Seu}{A Seurat object containing scRNA dataset with 'cell_cluster'
column.}

\item{by}{A character specifying the name of the metadata column referencing
the clusters.}

\item{biological_replicate_col}{Optional. A column of the Seurat object
indicating the replicates or batches of the dataset in order to take in
account biological/technical noise. If NULL, will create random layers of
fake replicates.}

\item{assay}{Assay to use.}
}
\value{
A pseudo-bulk matrice of cluster spread by replicates / batches /
fake replicates.
}
\description{
Create Pseudo-Bulk Matrice from scRNA clusters & replicates
}
\examples{
if(requireNamespace("Seurat", quietly=TRUE)){
data("Seu", package = "IDclust")
mat <- create_pseudobulk_mat_Seu(Seu, by = "seurat_clusters")
}
}
