% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{differential_edgeR_pseudobulk_LRT}
\alias{differential_edgeR_pseudobulk_LRT}
\title{Pseudo-bulk differential analysis with edgeR (LRT)}
\usage{
differential_edgeR_pseudobulk_LRT(
  Seu,
  by = "cell_cluster",
  assay = "RNA",
  biological_replicate_col = NULL,
  logFC.th = log2(1.5),
  qval.th = 0.01,
  min.pct = 0.1
)
}
\arguments{
\item{Seu}{A Seurat object containing scRNA dataset with a metadata column
name matching the by parameter}

\item{by}{A character specifying the name of the metadata column referencing
the clusters.}

\item{assay}{Assay to use.}

\item{logFC.th}{A numeric specifying the log2 fold change of activation
above/below which a feature is considered as significantly differential.}

\item{qval.th}{A numeric specifying the adjusted p-value below
which a feature is considered as significantly differential.}

\item{min.pct}{Minimum percentage of cells to be active in the cells of the
cluster to consider a feature as potentially significantly differential.}
}
\value{
A data.frame containing the results of the differential analysis
}
\description{
Pseudo-bulk differential analysis with edgeR (LRT)
}
\details{
Concatenate single-cells into replicates by cluster in order to
create a 'pseudo-bulk' matrice of multiple replicates per cluster. If no
replicates are present, will assign replicates at random to create 3 replicates
per cluster. Conducts 'LRT' (likelihood ratio tests) edgeR tests to test.
See \code{\link[edgeR:glmfit]{edgeR::glmLRT()}}
}
\seealso{
See \code{\link[edgeR:glmfit]{edgeR::glmLRT()}}
}
