% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_clustering_scHistone.R
\name{find_differentiated_clusters_scEpigenomics}
\alias{find_differentiated_clusters_scEpigenomics}
\title{Find Differentiated Clusters scEpigenomics}
\usage{
find_differentiated_clusters_scEpigenomics(
  scExp,
  qval.th = 0.1,
  min.frac = 0.1,
  logFC.th = 0.5,
  limit = 5,
  limit_by_proportion = NULL,
  min_frac_cell_assigned = 0.1,
  cluster_of_origin = "Omega"
)
}
\arguments{
\item{scExp}{A SingleCellExperiment with 'cell_cluster' column filled with
cluster assignations.}

\item{qval.th}{A numeric specifying the adjusted p-value threshold above
which a feature is defined as significantly differential.}

\item{min.frac}{A numeric between 0 and 1 specifying the minimum fraction of
cells activated for a given feature to consider it as differential.}

\item{logFC.th}{A numeric specifying the log Fold-Change of activation
(\% cells activated) threshold above which a feature is defined as
significantly differential.}

\item{limit}{An integer specifying the minimum number of features required
for a subcluster to be called 'true' subcluster.}

\item{limit_by_proportion}{Optional. A data.frame containing 3 columns -
ncells, mean_n_differential, sd_n_differential - that reflect the number of
false positive expected for a given cluster size.
(See  \link{calculate_FDR_scEpigenomics}).}

\item{min_frac_cell_assigned}{A numeric between 0 and 1 specifying the
minimum percentage of the total cells in the SingleCellExperiment object that
needs to be assigned. If a lower proportion is assigned, all cells are
assigned to the cluster of origin.}

\item{cluster_of_origin}{A character specifying the name of the cluster of
origin that will be concatenated before the name of true subclusters.}
}
\value{
A list containing :
\itemize{
\item "diffmat_n" - A data.frame containing the number of differential regions
foun per cluster and the new assignations of the subclusters.
\item "res" - A data.frame containing the differential analysis.
\item "passing_min_pct_cell_assigned" - A boolean indicating if enough cells
were  assigned
}
}
\description{
Find Differentiated Clusters scEpigenomics
}
\details{
Find significantly differential features between the given set
of clusters (within the 'cell_cluster' column of the SingleCellExperiment).
For each cluster, if enough differences are found, mark the cluster as a
'true' subcluster and gives it the alias 'cluster_of_origin:cluster'.
The function will use by default ChromSCape::differential_activation function
to define differential features.
}
\examples{

}
