% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_clustering_scRNA_Seurat.R
\name{find_differentiated_clusters_scRNA}
\alias{find_differentiated_clusters_scRNA}
\title{Iterative Differential Clustering scEpigenomics}
\usage{
find_differentiated_clusters_scRNA(
  Seu,
  output_dir = "./",
  plotting = TRUE,
  logFC.th = log2(1.5),
  qval.th = 0.01,
  test.use = c("edgeR-LRT", "Seurat")[1],
  biological_replicate_col = NULL,
  min.pct = 0.1,
  resolution = 0.8,
  starting.resolution = 0.1,
  limit = 5,
  cluster_of_origin = "Initial",
  nPCA = 50,
  min.pct.cell_assigned = 0.75,
  color = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Seu}{A Seurat object containing scRNA dataset with 'seurat_clusters'
column.}

\item{logFC.th}{}

\item{qval.th}{}

\item{test.use}{}

\item{biological_replicate_col}{}

\item{min.pct}{}

\item{limit}{}

\item{cluster_of_origin}{}

\item{min.pct.cell_assigned}{}
}
\value{

}
\description{
Iterative Differential Clustering scEpigenomics
}
