% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_clustering_scRNA_Seurat.R
\name{iterative_differential_clustering_scRNA}
\alias{iterative_differential_clustering_scRNA}
\title{Iterative Differential Clustering scRNA}
\usage{
iterative_differential_clustering_scRNA(
  Seu,
  method = c("Seurat", "monocle")[1],
  starting_differential_summary_df,
  output_dir = "./",
  plotting = TRUE,
  saving = TRUE,
  nPCA = 50,
  nfeatures = 2000,
  logFC.th = log2(2),
  qval.th = 0.01,
  min.pct = 0.1,
  min.pct.cell_assigned = 0.25,
  limit = 10,
  test.use = "wilcox",
  k = 100,
  resolution = 0.5,
  biological_replicate_col = "Target"
)
}
\arguments{
\item{Seu}{}

\item{method}{}

\item{starting_differential_summary_df}{}

\item{output_dir}{}

\item{plotting}{}

\item{saving}{}

\item{nPCA}{}

\item{nfeatures}{}

\item{logFC.th}{}

\item{qval.th}{}

\item{min.pct}{}

\item{min.pct.cell_assigned}{}

\item{limit}{}

\item{test.use}{}

\item{k}{}

\item{resolution}{}

\item{biological_replicate_col}{}
}
\value{

}
\description{
Iterative Differential Clustering scRNA
}
