% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_network}
\alias{plot_cluster_network}
\alias{plot_cluster_network.default}
\alias{plot_cluster_network.Seurat}
\title{Plot Iterative Differential Clustering network}
\usage{
plot_cluster_network(object, ...)

\method{plot_cluster_network}{default}(
  object,
  differential_summary_df,
  color_by = "cell_cluster",
  cluster_col = "cell_cluster",
  colors = NULL,
  node_size_factor = 7.5,
  edge_size_factor = 1,
  threshold_to_define_feature_active = 1,
  max_distanceToTSS = 1000,
  gene_col = "Gene",
  function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
    flip.y = FALSE),
  ...
)

\method{plot_cluster_network}{Seurat}(
  object,
  differential_summary_df,
  color_by = "cell_cluster",
  cluster_col = "cell_cluster",
  colors = NULL,
  node_size_factor = 7.5,
  edge_size_factor = 1,
  threshold_to_define_feature_active = 2,
  function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
    flip.y = FALSE),
  assay = "RNA",
  ...
)
}
\arguments{
\item{object}{A Seurat object clustered with \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}}}

\item{...}{Additional parameters passed to the plot function.}

\item{differential_summary_df}{Optional. A data.frame of differential
analyses summary outputed by \code{\link[=iterative_differential_clustering]{iterative_differential_clustering()}} when
saving option is TRUE. Use to determine the width of the edges based on the
number of differential features of the given marker.}

\item{color_by}{A character specifying the column of the Seurat
to use for coloring the nodes.}

\item{cluster_col}{A character specifying the column of the
Seurat to use to store the iterative differential clusters.}

\item{colors}{A character vector of colors. If NULL, will take R default
color.}

\item{node_size_factor}{A numeric specifying  a multiplicator of the size of
the nodes.}

\item{edge_size_factor}{A numeric specifying a multiplicator of the
size of the edges.}

\item{threshold_to_define_feature_active}{If color_by is a gene, an integer
specifying the threshold above which a gene is considered as active in any
given cell.}

\item{max_distanceToTSS}{If color_by is a gene, the maximum distance to TSS
to consider a gene  linked to a region. Used only if "color_by" is a gene
name.}

\item{gene_col}{If color_by is a gene, a character specifying the column in
the rowData of the object}

\item{function_layout}{A function of g for the layout of the graph.}

\item{assay}{If color_by is a gene, the assay in which to retrieve the counts.}
}
\value{


A hierarchical network of cluster assignation:
\itemize{
\item Size of nodes reflects the number of cells
\item Width of edges reflects the number of differential features defining a cluster
\item Color of nodes reflects the repartition of cells according to 'color_by'
}

A hierarchical network of cluster assignation:
\itemize{
\item Size of nodes reflects the number of cells
\item Width of edges reflects the number of differential features defining a cluster
\item Color of nodes reflects the repartition of cells according to 'color_by'
}
}
\description{
Plot Iterative Differential Clustering network

Plot Iterative Differential Clustering network

Plot Iterative Differential Clustering network
}
\examples{
# Plotting of Seurat scRNA object (Paired-Tag)
if(requireNamespace("Seurat", quietly=TRUE)){

data("Seu", package = "IDclust")
data("IDC_summary_scRNA", package = "IDclust")

plot_cluster_network(
    object = Seu,
    differential_summary_df = IDC_summary_scRNA,
    color_by = "cell_cluster",
    cluster_col = "cell_cluster",
    colors = NULL,
    node_size_factor = 7.5,
    edge_size_factor = 1,
    function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
                                                         flip.y = FALSE)
)

# Plotting proportion of cells activating a specific gene in  Seurat scRNA 
# object (Paired-Tag)
plot_cluster_network(
    object = Seu,
    differential_summary_df = IDC_summary_scRNA,
    color_by = "Erbb4", # a gene contained in the Seu object
    threshold_to_define_feature_active = 2,
    assay = "RNA",
    cluster_col = "cell_cluster",
    colors = NULL,
    node_size_factor = 7.5,
    edge_size_factor = 1,
    function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
                                                         flip.y = FALSE)
)

}

# Clustering of scExp scH3K27ac object (Paired-Tag)
if(requireNamespace("ChromSCape", quietly=TRUE)){

data("scExp", package = "IDclust")
data("IDC_summary_scEpigenomics", package = "IDclust")

plot_cluster_network(
    object = scExp,
    differential_summary_df = IDC_summary_scEpigenomics,
    color_by = "cell_cluster",
    cluster_col = "cell_cluster",
    colors = NULL,
    node_size_factor = 7.5,
    edge_size_factor = 1,
     
    function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
                                                         flip.y = FALSE)
)

# Plotting proportion of cells activating a specific gene in  scExp scH3K27ac 
# object (Paired-Tag)
plot_cluster_network(
    object = scExp,
    differential_summary_df = IDC_summary_scEpigenomics,
    color_by = "Tcf4", # a gene contained in the scExp object
    threshold_to_define_feature_active = 1,
    gene_col = "Gene",
    max_distanceToTSS = 1000,
    cluster_col = "cell_cluster",
    colors = NULL,
    node_size_factor = 7.5,
    edge_size_factor = 1,
    function_layout = function(g) igraph::layout_as_tree(g, root = 1, circular = TRUE,
                                                         flip.y = FALSE)
)

}
}
