% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{top_differential_markers.SingleCellExperiment}
\alias{top_differential_markers.SingleCellExperiment}
\title{Retrieve Top differential markers}
\usage{
top_differential_markers.SingleCellExperiment(
  scExp,
  IDC_DA_list,
  top = 1,
  gene_col = "Gene",
  pseudogene_pattern = "Rik|Vmn|Gm|AW",
  distanceToTSS = 1000
)
}
\arguments{
\item{scExp}{}

\item{IDC_DA_list}{}

\item{gene_col}{}

\item{pseudogene_pattern}{}

\item{distanceToTSS}{}
}
\value{

}
\description{
Retrieve Top differential markers
}
